fix(web): un 0 en el formulario ya no se reemplaza por el valor por defecto (QA H-07)
parseInt(v) || defecto trataba el 0 como vacio: un 0% en una clase de la
mezcla se enviaba como el % por defecto (el formulario decia 'Suma 100%' y el
servidor rechazaba 164%), '0 expansiones' de Cercania pasaba a 5 y en
Optimizacion/A-B un costo o penalizacion 0 volvia al default. La flota
ademas truncaba decimales (descarga 2.5 s -> 2).

- WebConfigurator.numero(id, defecto): default solo si el campo esta vacio
  o no es numero. Usado en mezcla, radios de Cercania, Optimizacion y A/B.
- fleet-manager: capacidad y descarga con parseFloat.

Reprobado desde la web (QA-1.3): mezcla 100/0/0/0/0 llega exacta y las 568
tareas son de clase pequeno. Plan de QA actualizado (QA-1.1 a 1.3, H-06,
H-07); analizador con desglose por clase.

+++ web_prototype/static/web_configurator/fleet-manager.js
@@ -541,8 +541,10 @@ class FleetManager {
             const groups = groundContainer.querySelectorAll('.fleet-group');
             groups.forEach(group => {
                 const cantidad = parseInt(group.querySelector('.input-cantidad').value);
-                const capacidad = parseInt(group.querySelector('.input-capacidad').value);
-                const tiempoDescarga = parseInt(group.querySelector('.input-tiempo-descarga').value);
+                // H-07: parseFloat, no parseInt -- parseInt truncaba los
+                // decimales (una descarga de 2.5 s se enviaba como 2).
+                const capacidad = parseFloat(group.querySelector('.input-capacidad').value);
+                const tiempoDescarga = parseFloat(group.querySelector('.input-tiempo-descarga').value);
 
                 // Get work area priorities
                 const workAreaPriorities = {};
@@ -573,8 +575,10 @@ class FleetManager {
             const groups = forkliftContainer.querySelectorAll('.fleet-group');
             groups.forEach(group => {
                 const cantidad = parseInt(group.querySelector('.input-cantidad').value);
-                const capacidad = parseInt(group.querySelector('.input-capacidad').value);
-                const tiempoDescarga = parseInt(group.querySelector('.input-tiempo-descarga').value);
+                // H-07: parseFloat, no parseInt -- parseInt truncaba los
+                // decimales (una descarga de 2.5 s se enviaba como 2).
+                const capacidad = parseFloat(group.querySelector('.input-capacidad').value);
+                const tiempoDescarga = parseFloat(group.querySelector('.input-tiempo-descarga').value);
 
                 // Get work area priorities
                 const workAreaPriorities = {};
